ChartParser

Description

A chart-based semiring parser for path grammars (e.g. the PV grammar). Path grammars operate on "paths" consisting of nodes (slices) and edges (transitions), both of which can contain arbitrary content. Paths are elaborated through two operations, splitting transitions and spreading slices (plus freeze, which terminates generation on a transition).

The parser is polymorphic in the grammar as well as the contents of slices (path nodes) and transitions (path edges). The grammar to parse is definend in an "evaluator" (Eval) which provides completions for parsing the splits, spreads and freezes.

Parsing Interface

parse

Arguments

:: Parsable tr slc v

=> (TChart tr slc v -> Either (VChart tr slc v) [Slice slc] -> Int -> IO ())
logging function

-> Eval tr tr' slc slc' v
the grammar's evaluator

-> Path slc' tr'
the input path (from first to last slice, excluding Start and Stop)

-> IO v
the semiring value at the top

The main entrypoint to the parser. Expects an evaluator for the specific grammar and an input path. Returns the combined semiring value of all full derivations.

parseSize :: Parsable tr slc v => Eval tr tr' slc slc' v -> Path slc' tr' -> IO v

Parse a piece using the logSize logging function.

parseSilent :: Parsable tr slc v => Eval tr tr' slc slc' v -> Path slc' tr' -> IO v

Parse a piece without logging.

logSize :: TChart tr1 slc1 v1 -> Either (VChart tr2 slc2 v2) [Slice slc2] -> Int -> IO ()

A logging function that logs the sice of the charts at each level.

logTikz :: (Foldable t, Show slc, Show slc, Hashable tr, Eq slc) => TChart tr slc v -> Either (VChart tr slc v) (t (Slice slc)) -> Int -> IO ()

A logging function that emits the state of the chart in TikZ code at every level.

Charts

Basic Elements

data Slice slc

A slice during chart parsing. Besides the slice content (e.g., notes), it maintains indices to the first and last surface slice covered, as well as an ID that is used for matching compatible parents of a spread.

data Transition tr slc

A transition during chart parsing. Has pointers to the two slices it connects, a content (e.g., protovoice connections), and a flag indicating whether it is the second (right) parent of a spread.

transLen :: Transition e a -> Int

Returns the "length" of the transition in terms of surface slices covered.

data Item i v

A parsing item. Combines an intermediate value (e.g. a transition) with a semiring score.

type TItem tr slc v = Item (Transition tr slc) v

A transition item.

Transition Chart

type TContents tr slc v = HashMap (Transition tr slc, Maybe (LeftId Int), Maybe (RightId Int)) (Score v Int)

The contents of a transition chart (under a particular index). A mapping from transitions (with score ID constraints left and right) to (partial) semiring scores. This mapping usually contains all transition items that satisfy a certain criterion, irrespective of their position in the chart (which is encoded in the transitions themselves).

When new transition items are added, if the transition already exists in the chart (as the result of a different partial parse), the scores of the new and existing items are "added" (this also requires the score IDs to match).

data TChart tr slc v

A transition chart. Stores intermediate transition items redundantly under several indices:

by surface length

by left slice

by right slice

tcGetByLength :: TChart tr slc v -> Int -> [TItem tr slc v]

Returns all transition items with the same length.

Verticalization Chart

data Vert tr slc v

Represents the middle part of an incomplete unspread ("verticalization"). Expresses how the middle transition and the two child slices (vMiddle) are derived from the parent slice (vTop) using a spread operation (vOp).

Vert objects are stored in the VChart to record the intermediate steps of an unspread, which is found by first parsing the middle transition into the parent slice (generating a Vert) and then combining the Vert with the left and right child transitions to generate the left and right parent transitions, respectively.

data VChart tr slc v

A verticalization chart. Stores Vert objects at certain chart positions. To support efficient lookup of Vert objects from different indices, each Vert is redundantly stored in several hash maps, one for each index:

by surface length

by surface length (only left border of a Vert)

by left child slice ID and mid transition length

by right child ID

In addition, the VChart maintains IDs of new slices. (Every new slice is the parent of an unspread.)

vcGetByLength

Arguments

:: VChart tr slc v
the chart

-> Int
surface length of a middle transition

-> [Vert tr slc v]
all corresponding Vert objects

Returns all Vert objects in the VChart with the same length.

Constraint Aliases

type Parsable tr slc v = (Normal tr, Normal slc, Normal' v)

A summary constraint for transitions, slices, and semiring values

type Normal x = (Eq x, Ord x, Show x, Hashable x, NFData x)

An alias for common constraints on slices and transitions

type Normal' x = (Eq x, Show x, NFData x, Semiring x)

An alias for common constraints on semiring values