Display

Description

This module contains functions for visualizing derivations using LaTeX and TikZ. These functions are generic over slice and transition contents and are thus somewhat limited. For protovoice derivations, an alternative to plotting a derivation would be to export an Analysis to JSON and view it interactively in the protovoice annotation tool (not implemented yet).

Plotting happens in two steps. First, a the derivation is "replayed" using a (generic or grammar-specific) "player" to construct a DerivationGraph, which contains all graphical objects and their positions explicitly. The DerivationGraph can then be plotted using different backends (currently only TikZ, but a diagrams/SVG backed would be useful too).

Documentation

data DerivationGraph slc tr

A derivation graph. Contains the graphical objects of a derivation plot as well as intermediate information that is used during a replay.

Constructors

DGraph

Fields

dgNextId :: !Int

a counter for generating new IDs (used during replay)

dgSlices :: !(Set (DerivSlice slc))

the positioned slices of the derivation graph

dgTransitions :: !(Set (DerivTrans slc tr))

the positioned transitionn in the derivation graph

dgHoriEdges :: !(Set (DerivSlice slc, DerivSlice slc))

the "horizontalization" edges (connecting the parent slice of a spread to its children)

dgOpen :: ![DerivTrans slc tr]

the open transitions of the current surface

dgFrozen :: ![DerivTrans slc tr]

the frozen transitions of the current surface in reverse order

dgRoot :: ![DerivTrans slc tr]

the root transitions

data DerivSlice slc

A slice together with a depth and an ID

Constructors

DerivSlice

Fields

dslDepth :: !Int

dslId :: !Int

dslContent :: !(StartStop slc)

type DerivTrans slc tr = (DerivSlice slc, tr, DerivSlice slc)

A transition between two DerivSlices.

Replaying Derivations

replayDerivation

Arguments

:: (Foldable t, Ord slc, Ord tr)

=> DerivationPlayer s f h slc tr
the derivation player

-> t (Leftmost s f h)
the derivation

-> Either String (DerivationGraph slc tr)

Replay a derivation from ⋊——⋉.

replayDerivation'

Arguments

:: (Foldable t, Ord slc, Ord tr)

=> Path tr slc
the starting point of the derivation

-> DerivationPlayer s f h slc tr
the derivation player

-> t (Leftmost s f h)
the derivation

-> Either String (DerivationGraph slc tr)

Replay a derivation from n top-level transitions.

replayDerivationFull

Arguments

:: (Foldable t, Ord slc, Ord tr)

=> DerivationPlayer s f h slc tr
the derivation player

-> t (Leftmost s f h)
the derivation

-> Either String (DerivationGraph slc tr)

Replay a derivation from ⋊——⋉ and ensure that the dervation is complete (freezing all transitions). Return an error message if not.

unfoldDerivation

Arguments

:: (Ord slc, Ord tr)

=> DerivationPlayer s f h slc tr
the derivation player

-> [Leftmost s f h]
the derivation

-> [Either String (DerivationGraph slc tr)]

Replays a derivation from ⋊——⋉ and returns every intermediate derivation graph.

unfoldDerivation'

Arguments

:: (Ord slc, Ord tr)

=> Path tr slc
the starting point of the derivation

-> DerivationPlayer s f h slc tr
the derivation player

-> [Leftmost s f h]
the derivation

-> [Either String (DerivationGraph slc tr)]

Replays a derivation from n top-level transitions and returns every intermediate derivation graph.

Derivation Players

data DerivationPlayer s f h slc tr

A derivation player. Contains functions for replaying derivations of a particular grammar, i.e. for deriving child elements from parent elements.

Constructors

DerivationPlayer

Fields

dpTopTrans :: !tr

the grammars default starting transition for ⋊——⋉

dpSplit :: !(s -> tr -> Either String (tr, slc, tr))

replay a split operation

dpFreeze :: !(f -> tr -> Either String tr)

replay a freeze operation

dpSpread :: !(h -> tr -> slc -> tr -> Either String (tr, slc, tr, slc, tr))

replay a spread operation

derivationPlayerUnit :: DerivationPlayer s f h () ()

A derivation player that uses () for slice and transition contents. The actual derivation operations are ignored, so only the outer structure is produced.

derivationPlayerEmpty :: DerivationPlayer s f h Empty Empty

A derivation player that uses Empty for slice and transition content. The actual derivation operations are ignored, so only the outer structure is produced.

data Empty

A helper type that is like () but has a Show instance that returns the empty string.

Plotting Derivation Graphs with TikZ

tikzDerivationGraph

Arguments

:: (Eq slc, Eq tr)

=> (slc -> Text)
a function for displaying slice contents

-> (tr -> Text)
a function for displaying transitions contents

-> DerivationGraph slc tr
the derivation graph

-> Text

Convert a derivation graph into a series of TikZ commands.

tikzPic :: (Semigroup a, IsString a) => a -> a

Wraps TikZ commands in a tikzpicture environment.

tikzStandalone

Arguments

:: (Semigroup a, IsString a)

=> Bool
a flag for using the the varwidth option of standalone (needed for multiline content)

-> a
the document content

-> a

Wraps latex code (e.g. a tikzpicture) in a complete standalone document. This environment includes default styles for slices, transitions, and hori edges.

writeGraph :: (Show slc, Eq slc, Eq tr, Show tr) => FilePath -> DerivationGraph slc tr -> IO ()

Write a single derivation graph to a tex file.

writeGraphs :: (Show tr, Show slc, Eq slc, Eq tr) => FilePath -> [DerivationGraph slc tr] -> IO ()

Write a several derivation graphs to a tex file.

viewGraph :: (Eq slc, Eq tr, Show slc, Show tr) => FilePath -> DerivationGraph slc tr -> IO ()

Write a single derivation graph to a tex file and compile the file using pdflatex.

viewGraphs :: (Show tr, Show slc, Eq slc, Eq tr) => FilePath -> [DerivationGraph slc tr] -> IO ()

Write a several derivation graphs to a tex file and compile the file using pdflatex.