FileHandling

Description

This module contains functionality for reading chord and slice data from csv files, as output by 'preprocess.py'. This also contains functionality for writing the results of the parses to json files, to be read by 'results.py'

Documentation

type InputSlice ns = ([(ns, RightTied)], Bool)

Slice type for reading from csv

pathFromSlices :: forall ns o. Eval (Edges ns) [Edge ns] (Notes ns) [ns] o -> SliceWrapper (Notes ns) -> [InputSlice ns] -> Path (Maybe [Edge ns], Bool) (SliceWrapped (Notes ns))

Create a Path given a list of slices

concatResults :: String -> String -> [ChordLabel] -> [Value] -> Value

Concatenates all results for a given piece into an object, inlucuding the piece and corpus in the JSON value.

slicesFromFile' :: FilePath -> IO [(Slice', Bool)]

Loads slices from filepath

chordsFromFile :: FilePath -> IO [ChordLabel]

Loads chord annotations from filepath

writeJSONToFile :: ToJSON a => FilePath -> a -> IO ()

Write JSON value to the given file

writeResultsToJSON :: [Notes SPitch] -> [ChordLabel] -> Maybe [PVLeftmost SPitch] -> Accuracy -> LogLikelihood -> String -> Time -> Value

Coverts the results of a parsing algorithm to a JSON value

nullResultToJSON :: Show a => a -> Value

Used when the algorithm fails. Contains NANs in all fields.

splitSlicesIntoSegments :: forall ns o. Eval (Edges ns) [Edge ns] (Notes ns) [ns] o -> SliceWrapper (Notes ns) -> [InputSlice ns] -> [[InputSlice ns]]

Orphan instances

FromField RightTied

Instance details

Methods

parseField :: Field -> Parser RightTied

FromField Bool

Instance details

Methods

parseField :: Field -> Parser Bool