HeuristicParser

Documentation

newtype Slice ns

Constructors

Slice

Fields

sContent :: ns

data Trans es

Constructors

Trans

Fields

tContent :: !es

t2nd :: !Bool

tBoundary :: !Boundary

data SearchState es es' ns o

The state of the search between steps. Generally, the current reduction consists of frozen transitions between the ⋊ and the current location and open transitions between the current location and ⋉.

⋊==[1]==[2]==[3]——[4]——[5]——⋉
  └ frozen  ┘  | └   open  ┘
            midSlice (current position)

frozen:   ==[2]==[1]==
midSlice: [3]
open:     ——[4]——[5]——

This is the SSSemiOpen case: The slice at the current pointer ([3]) is represented as an individual slice (_ssMidSlice). The frozen part is represented by a Path of frozen transitions (tr') and slices (slc). in reverse direction, i.e. from midslice back to ⋊ (excluding ⋊). The open part is a Path of open transitions (tr) and slices (slc) in forward direction from midSlice up to ⋉.

There are two special cases. All transitions can be frozen (SSFrozen), in which case state only contains the backward Path of frozen transitions (excluding ⋊ and ⋉):

⋊==[1]==[2]==[3]==⋉
                   └ current position
represented as: ==[3]==[2]==[1]==

Or all transitions can be open (SSOpen), in which case the state is just the forward path of open transitions:

⋊——[1]——[2]——[3]——⋉
└ current position
represented as: ——[1]——[2]——[3]——

The open and semiopen case additionally have a list of operations in generative order.

Types: es' : Unparsed/frozen transitions es : Unfrozen transitions ns : Evaluated slice content o : Operation type

The state of the search between steps

Constructors

SSFrozen !(Path (Maybe es', Bool) (Slice ns))

SSSemiOpen

Fields

_ssFrozen :: !(Path (Maybe es', Bool) (Slice ns))

frozen transitions and slices from current point leftward

_ssMidSlice :: !(Slice ns)

the slice at the current posision between gsFrozen and gsOpen

_ssOpen :: !(Path (Trans es) (Slice ns))

non-frozen transitions and slices from current point rightward

_ssDeriv :: ![o]

derivation from current reduction to original surface

SSOpen !(Path (Trans es) (Slice ns)) ![o]

showFrozen :: Show slc => Path (Maybe es', Bool) slc -> String

Helper function for showing the frozen part of a piece.

showOpen :: Show slc => Path (Trans es') slc -> String

Helper function for showing the open part of a piece.

exploreStates :: forall es es' ns ns' s f h. (Show s, Show f, Show h, Show es, Show es', Show ns, Show ns') => Eval es es' ns ns' (Leftmost s f h) -> SearchState es es' ns (Leftmost s f h) -> [SearchState es es' ns (Leftmost s f h)]

Returns a list of possible next states given the current state.

type Boundary = Bool

heursiticSearchGoalTest :: SearchState es es' ns o -> Bool

printPathFromState :: (Show es, Show ns) => SearchState es es' ns o -> String

getOpsFromState :: SearchState es es' ns o -> [o]

getPathLengthFromState :: SearchState es es' ns o -> Int

getPathFromState :: SearchState es es' ns o -> Maybe (Path es ns)

Parsing Actions

data ActionSingle ns tr s f

A parsing action (reduction step) with a single parent transition. Combines the parent elements with a single-transition derivation operation.

Constructors

ActionSingle

Fields

(StartStop ns, Trans tr, StartStop ns)

parent transition (and adjacent slices)

(LeftmostSingle s f)

single-transition operation

data ActionDouble ns tr s f h

A parsing action (reduction step) with two parent transitions. Combines the parent elements with a double-transition derivation operation.

Constructors

ActionDouble

Fields

(StartStop ns, Trans tr, ns, Trans tr, StartStop ns)

parent transitions and slice

(LeftmostDouble s f h)

double-transition operation

type Action slc tr s f h = Either (ActionSingle slc tr s f) (ActionDouble slc tr s f h)

An alias that combines ActionSingle and ActionDouble, representing all possible reduction steps.