PVGrammar

Description

This module contains common datatypes and functions specific to the protovoice grammar. In a protovoice derivations, slices are multisets of notes while transitions contain connections between these notes.

Code that is specific to parsing can be found in PVGrammar.Parse, while generative code is located in PVGrammar.Generate.

Inner Structure Types

Slices: Notes

newtype Notes n

The content type of slices in the protovoice model. Contains a multiset of pitches, representing the notes in a slice.

Constructors

Notes (MultiSet n)

innerNotes :: StartStop (Notes n) -> [StartStop n]

Return the notes or start/stop symbols inside a slice. This is useful to get all objects that an Edge can connect to.

Transitions: Sets of Obligatory Edges

Transitions contain two kinds of edges, regular edges and passing edges.

data Edges n

The content type of transitions in the protovoice model. Contains a multiset of regular edges and a multiset of passing edges. The represented edges are those that are definitely used later on. Edges that are not used are dropped before creating a child transition. A transition that contains passing edges cannot be frozen.

Constructors

Edges

Fields

edgesReg :: !(HashSet (Edge n))

regular edges

edgesPass :: !(MultiSet (InnerEdge n))

passing edges

topEdges :: Hashable n => Edges n

The starting transition of a derivation (⋊——⋉).

type Edge n = (StartStop n, StartStop n)

A proto-voice edge between two nodes (i.e. notes or start/stop symbols).

type InnerEdge n = (n, n)

A proto-voice edge between two notes (excluding start/stop symbols).

Generative Operations

Freeze

data Freeze

Represents a freeze operation. Since this just ties all remaining edges (which must all be repetitions) no decisions have to be encoded.

Constructors

FreezeOp

Split

data Split n

Encodes the decisions made in a split operation. Contains a list of elaborations for every parent edge and note. Each elaboration contains the child pitch, and the corresponding ornament. For every produced edge, a decisions is made whether to keep it or not.

Constructors

SplitOp

Fields

splitReg :: !(Map (Edge n) [(n, DoubleOrnament)])

Maps every regular edge to a list of ornamentations.

splitPass :: !(Map (InnerEdge n) [(n, PassingOrnament)])

Maps every passing edge to a passing tone. Since every passing edge is elaborated exactly once but there can be several instances of the same edge in a transition, the "same" edge can be elaborated with several passing notes, one for each instance of the edge.

fromLeft :: !(Map n [(n, RightOrnament)])

Maps notes from the left parent slice to lists of ornamentations.

fromRight :: !(Map n [(n, LeftOrnament)])

Maps notes from the right parent slice to lists of ornamentations.

keepLeft :: !(HashSet (Edge n))

The set of regular edges to keep in the left child transition.

keepRight :: !(HashSet (Edge n))

The set of regular edges to keep in the right child transition.

passLeft :: !(MultiSet (InnerEdge n))

Contains the new passing edges introduced in the left child transition (excluding those passed down from the parent transition).

passRight :: !(MultiSet (InnerEdge n))

Contains the new passing edges introduced in the right child transition (excluding those passed down from the parent transition).

data DoubleOrnament

Two-sided ornament types (two parents).

Constructors

FullNeighbor
a full neighbor note

FullRepeat
a repetition of both parents (which have the same pitch)

LeftRepeatOfRight
a repetition of the right parent

RightRepeatOfLeft
a repetitions of the left parent

RootNote
a note inserted at the top of the piece (between ⋊ and ⋉)

isRepetitionOnLeft :: DoubleOrnament -> Bool

Returns True if the child repeats the left parent

isRepetitionOnRight :: DoubleOrnament -> Bool

Returns True if the child repeats the right parent

data PassingOrnament

Types of passing notes (two parents).

Constructors

PassingMid
a connecting passing note (step to both parents)

PassingLeft
a step from the left parent

PassingRight
a step from the right parent

data LeftOrnament

Types of single-sided ornaments left of the parent (child-parent)

[ ] [p]
    /
  [c]

Constructors

LeftNeighbor
an incomplete left neighbor

LeftRepeat
an incomplete left repetition

data RightOrnament

Types of single-sided ornaments right of the parent (parent--child).

[p] [ ]
  \
  [c]

Constructors

RightNeighbor
an incomplete right neighbor

RightRepeat
an incomplete right repetition

Spread

data Spread n

Represents a spread operation. Records for every pitch how it is distributed (see SpreadDirection). The resulting edges (repetitions and passing edges) are represented in a child transition.

Constructors

SpreadOp !(HashMap n SpreadDirection) !(Edges n)

data SpreadDirection

Encodes the distribution of a pitch in a spread.

All instances of a pitch must be either moved completely to the left or the right (or both). In addition, some instances may be repeated on the other side. The difference is indicated by the field of the ToLeft and ToRight constructors. For example, ToLeft 3 indicates that out of n instances, all n are moved to the left and n-3 are replicated on the right.

Constructors

ToLeft !Int
all to the left, n fewer to the right

ToRight !Int
all to the right, n fewer to the left

ToBoth
all to both

Derivations

type PVLeftmost n = Leftmost (Split n) Freeze (Spread n)

Leftmost specialized to the split, freeze, and spread operations of the grammar.

type PVAnalysis n = Analysis (Split n) Freeze (Spread n) (Edges n) (Notes n)

An Analysis specialized to PV types.

analysisTraversePitch :: (Applicative f, Eq n', Hashable n', Ord n') => (n -> f n') -> PVAnalysis n -> f (PVAnalysis n')

Apply an applicative action to all pitches in an analysis.

analysisMapPitch :: (Eq n', Hashable n', Ord n') => (n -> n') -> PVAnalysis n -> PVAnalysis n'

Map a function over all pitches in an analysis.

Loading Files

loadAnalysis :: FilePath -> IO (Either String (PVAnalysis SPitch))

Loads an analysis from a JSON file (as exported by the annotation tool).

loadAnalysis' :: FilePath -> IO (Either String (PVAnalysis SPC))

Loads an analysis from a JSON file (as exported by the annotation tool). Converts all pitches to pitch classes.

slicesFromFile :: FilePath -> IO [[(SPitch, RightTied)]]

Loads a MusicXML file and returns a list of salami slices. Each note is expressed as a pitch and a flag that indicates whether the note continues in the next slice.

slicesToPath :: (Interval i, Ord i, Eq i) => [[(Pitch i, RightTied)]] -> Path [Pitch i] [Edge (Pitch i)]

Converts salami slices (as returned by slicesFromFile) to a path as expected by parsers.

loadSurface :: FilePath -> IO (Path [Pitch SInterval] [Edge (Pitch SInterval)])

Loads a MusicXML File and returns a surface path as input to parsers.

loadSurface'

Arguments

:: FilePath
path to a MusicXML file

-> Int
the first slice to include (starting at 0)

-> Int
the last slice to include

-> IO (Path [Pitch SInterval] [Edge (Pitch SInterval)])

Loads a MusicXML File and returns a surface path of the given range of slices.