Probability

Documentation

multinomialLogProb :: Vector Double -> Vector Double -> Double

Computes the multinomial log probability density function parameterised by prob, evaluated at the vector xs. The sum of the observations for all categories must add up to the number of samples, n. The probabilties must be NORMALISED!

sampleR :: StatefulGen g IO => g -> Int -> [a] -> IO [a]

categoricalLogProb :: Int -> Vector Double -> Maybe Double

pickRandomChoice :: StatefulGen g IO => g -> [slc] -> MaybeT IO slc

pickReservoir10 :: StatefulGen g IO => g -> [slc] -> MaybeT IO [slc]

pickRandomN :: StatefulGen g m => g -> Int -> [a] -> m [a]

collectRandomSample :: StatefulGen g IO => g -> MaybeT IO [a] -> IO [a]

collectRandomChoice :: StatefulGen g IO => g -> MaybeT IO a -> IO [a]

mixture :: Maybe (Vector Double) -> Maybe (Vector Double) -> Maybe (Vector Double)

data Prob a

Sample n elements from a list sample :: Show a => Int -> [a] -> Reservoir a [a] sample n xs = do let len = length xs let updateReservoir x = do ys <- get i <- lift $ randomRIO (0, len) if i < n then put (take i ys ++ [x] ++ drop (i + 1) ys) else return () liftIO $ putStrLn "Reservoir sampling..." liftIO $ putStrLn $ "List length: " ++ show len liftIO $ putStrLn $ "Reservoir size: " ++ show n liftIO $ putStrLn "Reservoir:" liftIO $ putStrLn "---------" put (take n xs) forM_ (drop n xs) $ x -> do updateReservoir x get >>= liftIO . putStrLn . show get

data RandomChoice a

testProb :: IO ()

poissonSample :: StatefulGen g IO => g -> Double -> IO Int

sample :: RandomChoice a -> StdGen -> [a] -> Prob a

reservoirSample :: StatefulGen g IO => g -> Int -> [a] -> IO [a]

pickRandom :: StatefulGen g m => g -> [a] -> m (Maybe a)