RandomChoiceSearch

Documentation

data RandomSearch s

Constructors

RandomSearch

Fields

rsHead :: s

rsGoal :: s -> Bool

randomSearchInit :: s -> (s -> Bool) -> RandomSearch s

randomChoiceSearch :: Show state => state -> (state -> ExceptT String IO [state]) -> (state -> Bool) -> (state -> String) -> ExceptT String IO state

Entry point to the search algorithm Picks random choice from a split or spread operation Skewed towards splits.

pickRandom :: StatefulGen g m => g -> [slc] -> m (Maybe slc)