HeuristicParser

Description

A parser for the protovoice model, representials partial reductions as a search state that can be traversed by applying operations such as Unspread, Unfreeze and Unsplit.

Documentation

newtype Slice ns

Slice datatype for parsing

Constructors

Slice

Fields

sContent :: ns

Content of the Slice

data SliceWrapped ns

Data type for a slice "wrapped" with additional information. Contains the content alongside the guess of the most likely chord label, and our confidence in that label.

Constructors

SliceWrapped

Fields

sWContent :: ns

sLbl :: ChordLabel

sLblProb :: Double

newtype SliceWrapper ns

Type for functions that wrap slices with additional information

Constructors

SliceWrapper

Fields

wrapSlice :: ns -> SliceWrapped ns

Converts a slice into a wrapped slice, giving additional information such as a guess of the chord it represents.

sliceWrapper :: SliceWrapper (Notes SPitch)

Returns a SliceWrapper that guesses the most likely chord for a slice

idWrapper :: SliceWrapper ns

data Trans es

Transition wrapper datatype. Wraps a transition with a flag indicating whether or not the transition is the 2nd right parent of a spread, and another flag marking whether or not this transition is at a boundary between segments

Constructors

Trans

Fields

tContent :: !es

The value inside the transition, i.e. protovoice edges

t2nd :: !Bool

Marks if this transition is the 2nd right parent of a spread

tBoundary :: !Boundary

Marks whether or not this transition is at a boundary between two segments

type Boundary = Bool

Boolean type synonym representing if a transition is a boundary between two segments

data SearchState es es' ns o

The state of the search between steps. Generally, the current reduction consists of frozen transitions between the ⋊ and the current location and open transitions between the current location and ⋉.

⋊==[1]==[2]==[3]——[4]——[5]——⋉
  └ frozen  ┘  | └   open  ┘
            midSlice (current position)

frozen:   ==[2]==[1]==
midSlice: [3]
open:     ——[4]——[5]——

This is the SSSemiOpen case: The slice at the current pointer ([3]) is represented as an individual slice (_ssMidSlice). The frozen part is represented by a Path of frozen transitions (tr') and slices (slc). in reverse direction, i.e. from midslice back to ⋊ (excluding ⋊). The open part is a Path of open transitions (tr) and slices (slc) in forward direction from midSlice up to ⋉.

There are two special cases. All transitions can be frozen (SSFrozen), in which case state only contains the backward Path of frozen transitions (excluding ⋊ and ⋉):

⋊==[1]==[2]==[3]==⋉
                   └ current position
represented as: ==[3]==[2]==[1]==

Or all transitions can be open (SSOpen), in which case the state is just the forward path of open transitions:

⋊——[1]——[2]——[3]——⋉
└ current position
represented as: ——[1]——[2]——[3]——

The open and semiopen case additionally have a list of operations in generative order.

Types: es' : Unparsed/frozen transitions es : Unfrozen transitions ns : Evaluated slice content o : Operation type

Constructors

SSFrozen !(Path (Maybe es', Bool) (SliceWrapped ns))

SSSemiOpen

Fields

_ssFrozen :: !(Path (Maybe es', Bool) (SliceWrapped ns))

frozen transitions and slices from current point leftward

_ssMidSlice :: !(SliceWrapped ns)

the slice at the current posision between gsFrozen and gsOpen

_ssOpen :: !(Path (Trans es) (SliceWrapped ns))

non-frozen transitions and slices from current point rightward

_ssDeriv :: ![o]

derivation from current reduction to original surface

SSOpen !(Path (Trans es) (SliceWrapped ns)) ![o]

type SearchState' = SearchState (Edges SPitch) [Edge SPitch] (Notes SPitch) (Leftmost (Split SPitch) Freeze (Spread SPitch))

data ActionSingle ns tr s f

A parsing action (reduction step) with a single parent transition. Combines the parent elements with a single-transition derivation operation.

Constructors

ActionSingle

Fields

(StartStop ns, Trans tr, StartStop ns)

parent transition (and adjacent slices)

(LeftmostSingle s f)

single-transition operation

data ActionDouble ns tr s f h

A parsing action (reduction step) with two parent transitions. Combines the parent elements with a double-transition derivation operation.

Constructors

ActionDouble

Fields

(StartStop ns, Trans tr, ns, Trans tr, StartStop ns)

parent transitions and slice

(LeftmostDouble s f h)

double-transition operation

exploreStates

Arguments

:: forall es es' ns ns' s f h. (Show s, Show f, Show h, Show es, Show es', Show ns, Show ns')

=> SliceWrapper ns

-> Eval es es' ns ns' (Leftmost s f h)

-> SearchState es es' ns (Leftmost s f h)

-> ExceptT String IO [SearchState es es' ns (Leftmost s f h)]
Either a error message or a list of possible next states

Returns a list of possible next states given the current state. Works by applying the inverse of the following operations, as returned by the protovoiceEvaluator. > > freeze left: split left: split right: spread: > ...=[]——[]——[]—... ...=[]——[]——[]—... ...=[]——[]——[]—... ...=[]——[]——[]—... > ...=[]==[]——[]—... > [] [] []——[]

goalTest :: [ChordLabel] -> SearchState es es' ns o -> Bool

Returns True if the parse is complete and has been reduced to only one slice per segment

goalTestSBS :: SearchState es es' ns o -> Bool

Returns True if the parse is complete, and has been reduced to only one slice

heursiticSearchGoalTest :: SearchState es es' ns o -> Bool

Goal test for a search: Only complete once there is a single slice per segment

getPathFromState :: SearchState es es' ns o -> Maybe (Path es ns)

getPathLengthFromState :: SearchState es es' ns o -> Int

getPathFromState' :: SearchState es es' ns o -> Maybe (Path es (SliceWrapped ns))

getSlicesFromState :: SearchState es es' ns o -> Maybe [SliceWrapped ns]

getOpFromState :: SearchState es es' ns o -> Maybe o

Returns the latest operation that has been applied, or Nothing if no operations have been applied

getOpsFromState :: SearchState es es' ns o -> [o]

Returns a list of all operations that have been taken up to the state given

chordAccuracy :: (Eq a, Show a) => [a] -> [a] -> Double

Calculate a naive accuracy metric for two lists using equality

guessChords :: [Notes SPitch] -> [ChordLabel]

Returns the most likely chord labels for each input group of notes

showOp :: (Show s, Show f, Show h) => [Leftmost s f h] -> String